+++ packages/dumb-durable-websocket/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ...
+## [0.0.5] - 2023-04-10
+
+### Changed
+
+- Changed `connect` to `accept`
+- Changed `accept` to return the other WebSocket instead of the response
+
 ## [0.0.4] - 2023-04-10
 
 ### Changed